feat: routed a concurrent build after its scan, dropping repartition().
A concurrent build cannot route during its scan, so a demux pass over
the segments it wrote routes them afterwards. That covers the case the
SQL function existed for, so it goes before it ships.
