+++ docs/changelog/unreleased/6146.feature.mdx
@@ -2,4 +2,4 @@
 header: features
 ---
 
-Added `paradedb.repartition()`, which routes an index's segments into the partitions its `partition_by` option asks for. An index that was created before its rows records no partition boundaries, so a range-partitioned join has nothing to cut on until the index is repartitioned or rebuilt.
+An index with `partition_by` now keeps its partitioned layout without a rebuild. A merge whose inputs carry no partition routes them, so rows that arrive through `INSERT` regain a partition on the merges that already run. `CREATE INDEX CONCURRENTLY` and `REINDEX CONCURRENTLY` route their segments in a pass after the build, so a concurrently built index no longer starts unpartitioned.
