--- pg_search/sql/unreleased/6146.repartition.sql
@@ -1,6 +1,0 @@
-CREATE  FUNCTION "repartition"(
-	"index" regclass /* PgRelation */
-) RETURNS bigint /* anyhow :: Result < i64 > */
-STRICT
-LANGUAGE c /* Rust */
-AS 'MODULE_PATHNAME', 'repartition_wrapper';
