+++ pg_search/tests/pg_regress/expected/partitioned_build_concurrently.out
@@ -1,0 +1,97 @@
+-- =====================================================================
+-- A `partition_by` index built with CREATE INDEX CONCURRENTLY. The scan
+-- of a concurrent build cannot route, so a pass after it rewrites the
+-- segments into partitions, and a range-partitioned join cuts on the
+-- split points it recorded. Results must match the serial baseline.
+-- =====================================================================
+CREATE EXTENSION IF NOT EXISTS pg_search;
+SET paradedb.enable_join_custom_scan TO on;
+SET paradedb.enable_range_partitioned_join TO on;
+SET max_parallel_workers TO 8;
+SET min_parallel_table_scan_size TO 0;
+SET parallel_setup_cost TO 0;
+SET parallel_tuple_cost TO 0;
+SET max_parallel_maintenance_workers TO 0;
+CREATE TABLE cic_users (id bigserial PRIMARY KEY, display_name text, about_me text);
+CREATE TABLE cic_posts (id bigserial PRIMARY KEY, owner_user_id bigint, title text, body text);
+INSERT INTO cic_users (display_name, about_me)
+SELECT 'user_' || g, repeat('a', 900) || g FROM generate_series(1, 20000) g;
+INSERT INTO cic_posts (owner_user_id, title, body)
+SELECT 1 + ((g * 7919) % 20000), CASE WHEN g % 3 = 0 THEN 'error in build ' ELSE 'note ' END || g, repeat('b', 900) || g
+FROM generate_series(1, 20000) g;
+CREATE INDEX CONCURRENTLY cic_users_idx ON cic_users USING paradedb (id, display_name)
+WITH (key_field = 'id', partition_by = 'id', target_segment_count = 4,
+      text_fields = '{"display_name": {"tokenizer": {"type": "keyword"}, "fast": true}}');
+WARNING:  only 0 parallel workers were available for index build
+CREATE INDEX CONCURRENTLY cic_posts_idx ON cic_posts USING paradedb (id, owner_user_id, title)
+WITH (key_field = 'id', partition_by = 'owner_user_id', target_segment_count = 4,
+      numeric_fields = '{"owner_user_id": {"fast": true}}');
+WARNING:  only 0 parallel workers were available for index build
+SELECT relname, count(*) AS segments
+FROM (SELECT 'cic_users_idx' AS relname FROM paradedb.index_info('cic_users_idx')
+      UNION ALL SELECT 'cic_posts_idx' FROM paradedb.index_info('cic_posts_idx')) s
+GROUP BY relname ORDER BY relname;
+    relname    | segments 
+---------------+----------
+ cic_posts_idx |        4
+ cic_users_idx |        4
+(2 rows)
+
+-- =====================================================================
+-- Serial baseline.
+-- =====================================================================
+SET max_parallel_workers_per_gather TO 0;
+SELECT count(*)
+FROM cic_users u JOIN cic_posts p ON u.id = p.owner_user_id
+WHERE u.id @@@ pdb.all() AND p.title @@@ 'error';
+ count 
+-------
+  6666
+(1 row)
+
+-- =====================================================================
+-- MPP: the scans show the boundaries the post-build pass recorded, and
+-- the rows match.
+-- =====================================================================
+SET max_parallel_workers_per_gather TO 3;
+EXPLAIN (COSTS OFF, VERBOSE, TIMING OFF)
+SELECT count(*)
+FROM cic_users u JOIN cic_posts p ON u.id = p.owner_user_id
+WHERE u.id @@@ pdb.all() AND p.title @@@ 'error';
+                                                                                                            QUERY PLAN                                                                                                             
+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+ Custom Scan (ParadeDB Aggregate Scan)
+   Output: pdb.agg_fn('COUNT(*)'::text)
+   Backend: DataFusion
+   Indexes: cic_users_idx (u), cic_posts_idx (p)
+   Aggregates: COUNT(*)
+   DataFusion Physical Plan: 
+     : ┌───── DistributedExec
+     : │ AggregateExec: mode=Final, gby=[], aggr=[count(1) as agg_0]
+     : │   CoalescePartitionsExec
+     : │     [Stage 1] => NetworkCoalesceExec: output_partitions=3, input_tasks=3
+     : └──────────────────────────────────────────────────
+     :   ┌───── Stage 1 ── tasks=3, partitions=3
+     :   │ AggregateExec: mode=Partial, gby=[], aggr=[count(1) as agg_0]
+     :   │   HashJoinExec: mode=Partitioned, join_type=Inner, on=[(owner_user_id@0, id@0)], projection=[]
+     :   │     DistributedLeafExec:
+     :   │       t0: PgSearchScan: table=p, segments=4, partition=owner_user_id[-∞..10001), query={"with_index":{"query":{"parse_with_field":{"field":"title","query_string":"error","lenient":null,"conjunction_mode":null}}}}
+     :   │       t1: PgSearchScan: table=p, segments=4, partition=owner_user_id[10001..15001), query={"with_index":{"query":{"parse_with_field":{"field":"title","query_string":"error","lenient":null,"conjunction_mode":null}}}}
+     :   │       t2: PgSearchScan: table=p, segments=4, partition=owner_user_id[15001..∞), query={"with_index":{"query":{"parse_with_field":{"field":"title","query_string":"error","lenient":null,"conjunction_mode":null}}}}
+     :   │     DistributedLeafExec:
+     :   │       t0: PgSearchScan: table=u, segments=4, partition=id[-∞..10001), dynamic_filters=1, query={"with_index":{"query":{"all":{"field":"id"}}}}
+     :   │       t1: PgSearchScan: table=u, segments=4, partition=id[10001..15001), dynamic_filters=1, query={"with_index":{"query":{"all":{"field":"id"}}}}
+     :   │       t2: PgSearchScan: table=u, segments=4, partition=id[15001..∞), dynamic_filters=1, query={"with_index":{"query":{"all":{"field":"id"}}}}
+     :   └──────────────────────────────────────────────────
+(23 rows)
+
+SELECT count(*)
+FROM cic_users u JOIN cic_posts p ON u.id = p.owner_user_id
+WHERE u.id @@@ pdb.all() AND p.title @@@ 'error';
+ count 
+-------
+  6666
+(1 row)
+
+DROP TABLE cic_posts;
+DROP TABLE cic_users;
